Add support for global constants to constant folder
Global constants like enums are represented in the AST as identifiers. These
identifiers can be looked up in the AST reference map. If the entry in the
reference map has a value associated with it, we can treat it as a constant
and attempt to turn it into a literal.

This reduces the size of the AST and unlocks more optimization
opportunities.